agents/frameworks/google_adk/pyproject.toml
[project]
name = "google-adk-local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=4.0.0",
    "google-adk[eval]>=1.17.0",
    "litellm<=1.76.3",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "ibmi-agent-sdk",
    "ibm-watsonx-ai>=1.4.1",
]

[tool.poetry.dependencies]
google-adk = { version = "*", extras = ["eval"] }

[tool.uv.sources]
ibmi-agent-sdk = { path = "../../packages/ibmi-agent-sdk", editable = true }